+++ src/base/asset_loader_zip.c
@@ -25,6 +25,7 @@
 #include "miniz/miniz.h"
 #include "miniz/miniz_zip.h"
 #include "base/asset_loader_zip.h"
+#include "base/assets_manager.h"
 
 typedef struct _asset_loader_zip_t {
   asset_loader_t asset_loader;
@@ -79,7 +80,10 @@ static void* miniz_awtk_realloc_func(void* opaque, void* address, size_t items, 
 
 static bool_t asset_loader_zip_exist(asset_loader_t* loader, const char* path) {
   asset_loader_zip_t* zip = (asset_loader_zip_t*)loader;
-  int file_index = mz_zip_reader_locate_file(&(zip->archive), path, NULL, 0);
+  const char* res_root = assets_manager_get_res_root(assets_manager());
+  uint32_t res_root_len = res_root == NULL ? 0 : strlen(res_root);
+  const char* p = path + ((res_root_len == 0) ? 0 : (res_root_len + 1));
+  int file_index = mz_zip_reader_locate_file(&(zip->archive), p, NULL, 0);
 
   return file_index >= 0;
 }
@@ -90,7 +94,10 @@ static asset_info_t* asset_loader_zip_load(asset_loader_t* loader, uint16_t type
   void* data = NULL;
   asset_info_t* info = NULL;
   asset_loader_zip_t* zip = (asset_loader_zip_t*)loader;
-  int file_index = mz_zip_reader_locate_file(&(zip->archive), path, NULL, 0);
+  const char* res_root = assets_manager_get_res_root(assets_manager());
+  uint32_t res_root_len = res_root == NULL ? 0 : strlen(res_root);
+  const char* p = path + ((res_root_len == 0) ? 0 : (res_root_len + 1));
+  int file_index = mz_zip_reader_locate_file(&(zip->archive), p, NULL, 0);
 
   if (file_index < 0) {
     return NULL;